Camada Controls: Adicionado a classe SearchUserMethods, para que a mesma receba os métodos de seu respectivo form.
