+++ QrAguas/Controls/Functions.cs
@@ -71,44 +71,6 @@ namespace QrAguas.Controls
 
         #endregion
 
-        #region Form SearchUser
-
-        public DataTable ProcurarUsuario(string nomeUsuario)
-        {
-            DataTable dataTable = new DataTable();
-
-            string queryProcurarUsuario = "SELECT U.ID_USUARIOS, U.NOME_USUARIO, T.DESCRICAO, U.DATA_CADASTRO, U.CADASTRADO_POR, U.ATIVO FROM USUARIOS U, TIPO_USUARIO T WHERE U.ID_TIPO_USUARIO = T.ID_TIPO_USUARIO AND U.NOME_USUARIO = @NOME_USUARIO AND U.ID_TIPO_USUARIO > @ID_TIPO_USUARIO";
-
-            
-            MySqlCommand command = new MySqlCommand(queryProcurarUsuario, AbrirBanco());
-
-            command.Parameters.AddWithValue("@NOME_USUARIO", nomeUsuario);
-            command.Parameters.AddWithValue("@ID_TIPO_USUARIO", Login.TipoUsuario);
-
-            MySqlDataReader reader = command.ExecuteReader();
-
-            dataTable.Load(reader);
-
-            return dataTable;
-
-        }
-
-        public bool AtivarDesativarUsuario(int idUsuario, bool ativo)
-        {
-            string queryAtivarDesativarUsuario = "UPDATE USUARIOS SET ATIVO = @ATIVO WHERE ID_USUARIOS = @ID_USUARIOS";
-
-            MySqlCommand command = new MySqlCommand(queryAtivarDesativarUsuario, AbrirBanco());
-
-            command.Parameters.AddWithValue("@ATIVO", ativo);
-            command.Parameters.AddWithValue("@ID_USUARIOS", idUsuario);
-
-            int rowCount = command.ExecuteNonQuery();
-
-            return rowCount != 0 ? true : false;
-        }
-
-        #endregion
-
         #region Form SellProduct
 
         public Product ProcurarProduto(int codigoProduto)
