+++ QrAguas/QrAguas.csproj
@@ -101,6 +101,7 @@
     <Compile Include="Controls\RegisterProductMethods.cs" />
     <Compile Include="Controls\RegisterProvidersMethods.cs" />
     <Compile Include="Controls\RegisterUserMethods.cs" />
+    <Compile Include="Controls\SearchUserMethods.cs" />
     <Compile Include="Controls\UpdateCategoryMethods.cs" />
     <Compile Include="Models\CepResponse.cs" />
     <Compile Include="Models\ICepApiService.cs" />
